HandWritting/OCR-master/backend/requirements.txt
numpy==1.19.2
Werkzeug==1.0.1
Flask==1.1.2
imutils==0.5.4
Flask_Cors==3.0.10
vietocr==0.3.2
requests==2.24.0
tensorflow==2.4.1
matplotlib==3.3.2
Keras==2.4.3
opencv_contrib_python==4.5.1.48
symspellpy==6.7.0
scipy==1.5.2
Pillow==8.2.0
scikit_learn==0.24.2